offre

+++ fr/univ-amu/src/java/main/fr/univamu/iut/Client/Client.java
@@ -58,17 +58,16 @@ public abstract class Client {
         return mesProduitsAchetes;
     }
 
-    public void validerOffre(){
-        if(offreAchat.isAccepter() && offreAchat.getMonCreateur().getMonComte().getSolde() > offreAchat.getMontant()){
-            offreAchat.getMonCreateur().acheterProduit(offreAchat.getProduitConcerne(), offreAchat.getProduitConcerne().getProprietaire());
-        }
-        else{
+    public void validerOffre() {
+        if (offreAchat.isAccepter() && offreAchat.getMonCreateur().getMonComte().getSolde() > offreAchat.getMontant()) {
+            offreAchat.getMonCreateur().acheterProduit(offreAchat.getProduitConcerne(), offreAchat.getProduitConcerne().getProprietaire(), offreAchat.getQuantite());
+        } else {
             System.out.println("Vous ne pouvez pas acheter ce produit...");
         }
     }
 
-    public void accepterOffre(boolean accepter){
-        if(accepter){
+    public void accepterOffre(boolean accepter) {
+        if (accepter) {
             offreAchat.setConforme(accepter);
 
         }
@@ -85,17 +84,26 @@ public abstract class Client {
     public void setNom(String nom) {
         this.nom = nom;
     }
-    public void acheterProduit(Produits produit, Client vendeur){
+
+    public void acheterProduit(Produits produit, Client vendeur, int quantite) {
+        if(quantite > produit.getQuantite()){
+            System.out.println("Pas assez de produits pour cette transaction");
+        }
+        else{
+            int newQuantite = produit.getQuantite() - quantite;
+            produit.setQuantite(newQuantite);
             mesProduitsAchetes.add(produit);
-            monComte.debiter(idClient, produit.getPrix());
-            produit.getProprietaire().getMonComte().crediter(produit.getProprietaire().getIdClient(), produit.getPrix());
+            monComte.debiter(idClient, produit.getPrix() * quantite);
+            produit.getProprietaire().getMonComte().crediter(produit.getProprietaire().getIdClient(), produit.getPrix() * quantite);
             produit.getProprietaire().supprimerProduit(produit);
             transaction = new Transaction(vendeur, this, produit);
             transaction.addTransaction(transaction);
+        }
+
 
     }
 
-   public String notifierClient(Client client){
+    public String notifierClient(Client client) {
         return "De nouveaux produits ont été mis en vente pour vous " + client.getNom();
     }
 
@@ -114,4 +122,6 @@ public abstract class Client {
     public void setOffreAchat(OffreAchat offreAchat) {
         this.offreAchat = offreAchat;
     }
+
+}
